requirement.txt
instagrapi>=2.2.0
python-dotenv>=1.0.0